model_settings/pyproject.toml
[project]
name = "model-settings"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-agents>=0.3.1",
    "python-dotenv>=1.1.1",
]
